Creation CashService for url api and CashInfo In CashWelcomeViewController new func updateCash

+++ P9_01_Le_Baluchon/Model/UserWeather.swift
@@ -20,7 +20,7 @@ struct WeatherElement: Decodable {
 }
 
 struct Main: Decodable {
-    let temp: Float
+    let temp: Double
 }
 
 enum CodingKeys: String, CodingKey {
